refactor(web): extract backtesting portfolio actions
